Add new connection management methods and POST route to add a New Connection
Added `CheckConnectionExists` and `AddNewConnection` methods in the `ClientHelper` class to manage connections in the database. Introduced a new POST route `/Clients/Connections` in the `ClientRoute` class to handle connection creation, including validation of client existence and connection uniqueness. Added a `Connection` class to represent the JSON object for the new route.